pyproject.toml
[project]
authors = [{name = "Haoyang Zhou", email = "me@haoyang.graphics"}]
dependencies = ["lightning>=2.5.5,<3", "tensorboard>=2.20.0,<3", "torchdiffeq>=0.2.5,<0.3", "jsonargparse[signatures]>=4.42.0,<5", "meshio>=5.3.5,<6", "h5py>=3.15.1,<4"]
name = "neural-ode"
requires-python = ">= 3.11"
version = "0.1.0"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pixi.pypi-dependencies]
neural_ode = { path = ".", editable = true }

[tool.pixi.tasks]

[tool.pixi.dependencies]
fenics-dolfinx = ">=0.10.0,<0.11"
mpich = ">=4.3.2,<5"
pyvista = ">=0.46.3,<0.47"

[tool.black]
line-length = 80
